backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-pptx==0.6.23
openai>=1.50.0
pydantic==2.6.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-core>=0.1.0
